# Conflicts:
#	frontend/locales/nl.json

+++ backend/research/mutations/DeleteStudyMutation.py
@@ -8,7 +8,7 @@ from main.models import MRPermission
 
 class DeleteStudyMutation(Mutation):
 
-    ok = Boolean()
+    ok = Boolean(required=True)
     errors = List(ErrorType)
 
     class Arguments:
@@ -28,7 +28,7 @@ class DeleteStudyMutation(Mutation):
                     Study.objects,
                     info,
                 )
-                .accessible_objects(info.context.user, MRPermission.EDIT)
+                .accessible_objects(info.context.user, MRPermission.DELETE)
                 .get(pk=id)
             )
         except Study.DoesNotExist:
@@ -36,7 +36,14 @@ class DeleteStudyMutation(Mutation):
                 field="id",
                 messages=["Study not found."],
             )
-            return cls(errors=[error])
+            return cls(errors=[error], ok=False)
 
-        study.delete()
+        # Soft delete if the study has ever been submitted.
+        if study.has_been_submitted:
+            study.is_deleted = True
+            study.save()
+        # Otherwise, do a regular (hard) delete.
+        else:
+            study.delete()
+
         return cls(ok=True)
